Refactor create-campaign form and validators for imageUrl field
Add mat-labels for better user experience and adjust form styles.

Remove Validators.required for the imageUrl field.

Update onSubmit method to check if an imageUrl is provided in the form value. If no imageUrl is provided, the default-image from the assets folder is assigned to the imageUrl field.

Adjust the image-url.validator to accommodate the changes in validation requirements.

Add class 'valid' or 'invalid' to the form button based on form validity.